Move timeframe config out of strategy
