+++ crypto_screener/domain/strategies/__init__.py
@@ -1,0 +1,5 @@
+from crypto_screener.domain.strategies.ppo import PpoStrategy
+
+__all__ = [
+    "PpoStrategy",
+]
